Improve readability by exctracting a UI.flash method
